Add files via upload
Scheme macros.

BeamSteeringMacros: Original set of macros, includes loop to rotate beam steering mirrors and collect data from screens along the beam path. 

BeamWaist_SampleTests: This code is to test the effect of the lenses on the focal length of the beam for a simplified setup that only consists on the lenses and quartz bar in a straight line.

BeamWaistTests: This code is to test the effect of the position lenses on the focal length of the beam for the full setup of the system. It collects irradiance screen data at 0 deg for a variety of L1 and L2 positions with a constant separation

GeometricDeadSpots: Code to collect info for Geometric Dead spots test with the fake laser source (rotating laser source instead of . Not complete: need to get phi/theta rotations to work. Attempts to be able to have the light be able to hit any point along the surface. Not used.

GeometricDeadSpotsDEADCenter: Code to collect info for Geometric Dead spots test with the fake laser source (rotating laser source instead of .Laser source always hits dead center of light pipe. Currently works. 

LaserPathScreens: ;This code finds the Ionization Intensity of a beam at a given angle. For rotating ONLY theta at paired mirror angles (no moving them separately), it collects the irradaiance screens along the beam path.

ThetaPhiRot: Code to run phi/theta rotations with the mirror setup and save data. Currently not working, still need to get theta and phi rots at same time.